+++ lib/hcloud/entities/floating_ip_prices.rb
@@ -3,6 +3,6 @@
 module HCloud
   class FloatingIPPrices < Entity
     attribute :type
-    attribute :prices, :price
+    attribute :prices, :price, array: true, default: -> { [] }
   end
 end